[Middleware] [Frontend] Google OAuth feature (#9)
* Updated: google oauth basic
currently ensure user can login and out with the oauth
ensure only 1 user is created / save to DB
attached token to the user object
need to work that in with express session storing into mongoDB
need to ensure can load stuff from user profile.

* Modified: google oauth import settings.

* Updated: persistent session.
added ability to save session to DB as well.

* Updated: reformatted code and new feature
reformat code such that we have dedicate place for authentication and routing and not all in server.js
index.js handle an extra profile selection to ensure the frontend can access the data.

* Modified: make google port flexible to user define server port.

* Patched: typo in Login page.

* Modified: uncomment strict mode.